added LocationService
